more classes

+++ CS3388_Asn2_0/Normal.h
@@ -4,6 +4,6 @@
 #include "opencv2/core/core.hpp"
 #include "opencv2/opencv.hpp"
 
-typedef cv::Point3d Normal;
+typedef cv::Point3f Normal;
 
 #endif
